Adds missing extensions files from git into project
