+++ Team4Bootcamp/Team4Bootcamp/Extensions/UIColor+Palette.swift
@@ -1,0 +1,15 @@
+//
+//  UIColor+Palette.swift
+//  Team4Bootcamp
+//
+//  Created by alessandra.l.pereira on 04/04/18.
+//  Copyright © 2018 alessandra.l.pereira. All rights reserved.
+//
+import UIKit
+
+extension UIColor {
+    static let primaryColor = UIColor(named: "PrimaryColor")
+    static let secondaryColor = UIColor(named: "SecondaryColor")
+    static let terciaryColor = UIColor(named: "TerciaryColor")
+    static let accentColor = UIColor(named: "AccentColor")
+}
