+++ Team4Bootcamp/Team4Bootcamp/Supporting Files/Assets.xcassets/Color Pallete/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
