+++ Team4Bootcamp/Team4Bootcamp/Supporting Files/Assets.xcassets/Color Pallete/PrimaryColor.colorset/Contents.json
@@ -1,0 +1,20 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "colors" : [
+    {
+      "idiom" : "universal",
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "red" : "0xF7",
+          "alpha" : "1.000",
+          "blue" : "0x5B",
+          "green" : "0xCE"
+        }
+      }
+    }
+  ]
+}
