+++ Team4Bootcamp/Team4Bootcamp/Supporting Files/Assets.xcassets/Color Pallete/SecondaryColor.colorset/Contents.json
@@ -1,0 +1,20 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "colors" : [
+    {
+      "idiom" : "universal",
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "red" : "0x2D",
+          "alpha" : "1.000",
+          "blue" : "0x47",
+          "green" : "0x30"
+        }
+      }
+    }
+  ]
+}
