+++ Team4Bootcamp/Team4Bootcamp/Supporting Files/Assets.xcassets/Color Pallete/TerciaryColor.colorset/Contents.json
@@ -1,0 +1,20 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "colors" : [
+    {
+      "idiom" : "universal",
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "red" : "0xD8",
+          "alpha" : "1.000",
+          "blue" : "0xD8",
+          "green" : "0xD8"
+        }
+      }
+    }
+  ]
+}
